Check if symlink for vimrc already exists before recreating

+++ install.sh
@@ -2,4 +2,5 @@
 
 DOTFILES_LOCATION="$( cd "$( dirname "${BASH_SOURCE[0]}" )" >/dev/null 2>&1 && pwd )"
 
-ln -s "$DOTFILES_LOCATION/.vimrc" "$HOME/.vimrc"
+# Add the vimrc file
+[ ! -f "$HOME/.vimrc" ] && ln -s "$DOTFILES_LOCATION/.vimrc" "$HOME/.vimrc"
